Start project over
I know I've already done a project where I use an Erlang node running
Cowboy to serve up animation instructions to a web page using
websockets. However, I can't find it, and this repository was basically
empty on GitHub, so I'm taking it over and starting from scratch.

So far I have a working Erlang Cowboy web server with a websocket, and a
web page that connects to it.