Re-organize, rename, add channel registry
Rename "server" to "channel":
Multiple websockets can join to the same channel so that multiple
browser windows can be part of the same animation. The ideas that you
might have a window for animating, a window for logging and a window for
setting animation parameters.

Rename "ws_handler" to "socket":
A "ws_handler" was to handle websockets, so "socket" made more sense.

Add a channel registry so that browsers can start and join channels. The
first browser window has to create the channel, which returns a name.
The names are currently just increasing integers starting from 1. Once a
channel is created, the first browser window can spawn other windows and
pass them the channel ID. Broswer windows can also ask for a list of
channels that currently exist.

The idea is that a "channel" is a many-to-many coordinator between
potentially multiple web sockets and multiple animation processes.
There's no reason to restrict an animation to a single process.

ws_anim_socket has a bunch of commands it can respond to and all
responses should be in JSON format. The browser window is responsible to
decode the JSON to figure out what to do: there will be a simple
protocol for passing back updated settings (maybe?), log messages and
animation instructions.

I'm not sure if the animation server is going to pass back updated
state, or if the server will send all draw calls. I don't want to fall
into the trap of thinking about what's feasible: this is just for fun
and it doesn't have to be performant. The server can just pass a bunch
of shapes to the web page and then the page can draw them. What I might
do is have an Erlang process that collects all draw calls and holds them
until a draw timer goes off: that way animation processes can post their
draw calls at whatever rate they want.

Oh, I also added a notes file so I can track what I'm thinking without
cluttering up source files with comments.