Add subscriptions
Sockets can subscribe to different events. So far log messages are just
given to all sockets by default and the only type that a socket might
want to subscribe to is the 'draw' type.

A browser window with a canvas will subscribe to the draw calls.

I still need to add a buffere to the channel so it can collect draw
calls before sending them to the socket in a batch. It might be better
to have the browser window do double buffering so it can draw to a
background buffer while rendering the current buffer. We'll see.

Sockets can check what types of messages they're subbed to.