Update to Erlang 27.0; get animation working
I'm now using the built-in JSON encoder since it's much easier than
building JSON manually. I was too lazy to include Jiffy before. (Dumb as
that sounds.)

Sometimes the web page just stops drawing more squares and I don't know
why. It stopped at 168 once. Not sure if that's relevant. It doesn't
stop at 168 squares each time.

I've purposely set the animator and the buffer out of sync so sometimes
the animator manages to make multiple draw calls for each frame.

It's stupid that in JavaScript you can't both destructure a parameter
and capture the paramter. (At least not that I or ChatGPT know of.)