WIP: separate browser window for animation ctrls
The idea is to have a separate browswer window that can control each of
the animations being rendered on the main browser window.

The controls are created and recreated based on messages from the
server. The messages are of type "control". Browser windows must
subscribe to the "control" message type in the channel.

There is a control message "clear" which should clear the screen. I've
read that assigning a new body element to the document in an HTML page
will clear the screen, so that's what I'm going to try first.

Since each new window needs to join the channel of the original window
and then subscribe to events, I'm opening the new window with the
channel in the query string. In order to do that, the channel has to
return the channel name when the channel is created, and it has to come
in an expected format. I've added a new "info" message type and
JavaScript in the main page to check for an info object containing the
"channel_name" property.

I want to keep returning log messages to the connected websockets, so
the Erlang socket module now expects a list of text messages to send
back to the socket after running the do/2 function.

I'm envisioning a bunch of control messages coming from the server to
any browser window subscribed to them any time the set of controls
changes, e.g. if a new animation is added. This doesn't include when any
control _setting_ changes, just the number & type of controls available.
This means the page should listen for "control" messages and render the
controls as they come in, and clear the screen if a "clear" control
message comes in.

Each control that the server sends needs to come with the current
setting so that all the settings displayed on the screen don't get reset
to their default values.

When the user changes the control settings the new settings should be
sent to the server to update the animations.

One issue I'll have to solve is coming up with a new name for any
animators that are added. The first control that gets sent is a list of
possible animators created by the channel, based on hard-coded knowledge
of what animators are available. If the user picks an animator, that
animator should start with a new, unique name so that controls can
reference that particular animator.

I'm not sure how I feel about having a bunch of animators with the same
module but unique settings: the settings page might get really busy,
really fast. However, I'm not a UX designer so I'm not sure if I'll even
try to solve that. I do also plan to have a text-based shell window
where the user can type commands. I don't think I want the user to have
to add something like 50 animators in order to do something like have a
bunch of circles bouncing around. Then again, I could just have an
animator that does "X circles" or "Between X and Y circles"; a valid
choice would be 1 circle, but you wouldn't have to create 50 animators
if all the circles are doing very similar things. I mean, I suppose lots
of different animators could allow for a count or range.