Get controls working
Not all the controls are finished, but controls are working.

I'm still just using the "square" animation, I've only created controls
for width, height and style, and only width can actually be set.

Every control needs to be spelled out in detail for the controls browser
window to create the input element, label and event listeners.

I've added a new command to the web socket and new decoder logic to the
channel to try an parse out the name of the animator, the name of the
field and the value. I had to change the animators list in the channel
state so that I could look up channels by name: the browser windows only
know the animator name attached to each control, not the Erlang pid, so
we look up the pid from the name.

The channel sends its controls first (clear and "pick an animator") and
then all the animators are told to send their controls. I'll need add
some styling or UI elements that identify different controls for
different animators; currently I'm using the animator name, which works
for now.

The controls window now disappears if the websocket closes so that I
don't have a bunch of controls windows hanging around. I couldn't seem
to access the new window to use it in a websocket event listener because
the browser said it wasn't initialized. I added a "load" event listener
on the new window and when it's loaded, _then_ I add the websocket
event listener that closes the controls window when the websocket
closes. I could have used a promise for this, I think, ... maybe.