Update objects in buffer instead of deleting
Keep all the buffered objects to draw in the buffer, which is a now an
ETS table. The animators update the objects in the buffer so that if the
channel and animators get out of sink, objects don't disappear off the
screen.

Previously the channel would delete objects from the buffer. If the
animator didn't get a new object into the buffer in time for the next
draw call, the screen would be cleared and that object wouldn't be
redrawn.

Now, if the animator doesn't update an object, the old version will
still be in the buffer to be redrawn in place.

Also, WIP on a new animator.

Update JS with tips from a stream viewer. (Thanks!)

Add missing animator.