New animator: draw circles rotating around
The new animator draws a circle that rotates around a point, then a
smaller circle that rotates around the first circle, then a third circle
that rotates around the second, and so on. One hundred lines are drawn
to trace out the last center position of the last circle.

Since the buffer is an ordered_set, if I re-use an ID I can effectively
create a queue of drawing objects. By taking the modulus of the current
frame by "max lines", I can keep looping back over existing IDs. Once an
ID is re-used it looks like the object is pulled off the queue and
replaced, but really, it's just updated.

Given max IDs of 5:

  Frame    Mod by max     Add starting ID of 10
  1        1 % 5 = 1      11
  2        2 % 5 = 2      12
  3        3 % 5 = 3      13
  4        4 % 5 = 4      14
  5        5 % 5 = 0      10
  6        6 % 5 = 1      11   <- overwrite object 11

Also remove the buffer 'type' argument, since I don't need it and don't
remember what it was for.