Freeze animations
Add buttons to freeze an animation; the animation is frozen by keeping
the frame count the same.

If I ever have animations that do not rely on the frame count then I'll
need to cache the last batch of draw calls and keep sending them.
So, as much as possible, all animations should either be based on the
frame, or last check if the frame has changed.