Don't show >1 control windows for one animation
I don't want all kinds of control windows created if the user keeps
clicking on the button for the particular animation.

If you set a property on an object in JavaScript, "in" will continue
working until you "delete" the property. Even if you set the property to
undefined, "in" will still return true.

e.g.

a = {b: 1};
'b' in a;         // true
a.b = undefined;
'b' in a;         // true
delete a.b;
'b' in a;         // false

I tried to attach a listener to the animation button in the main window,
from the new window, but it didn't seem to work.

i.e. if the main window is "main" and the new window is "new", in the
"new" load event listener I tried to add a listener for a button in
"main"; however, when the button in main was clicked, "new" didn't seem
to see the event.