Load and update filenames
Create a select list of files in the save/ directory to load.
Create a datalist-backed text field to show existing file names to save,
or allow entering a new one.

When a new file is saved, or when a websocket subscribes to "info", all
filenames are sent to the websocket. When the web page receives the
filenames, it updates both the load and save controls. A select and a
datalist cannot share the same option: if you add an option to one
(i.e. with appendChild) and then add it to the other, it will move, not be
duplicated. Note, this means the same DOM object, not just an option
with the same text.

It _seemed_ like childNodes.forEach((c) => c.remove()) didn't work, and
an AI suggested a while(firstChild){...} method, which worked.

The filenames are also sent when the channel first starts, but it's
likely that the websocket hasn't had a chance to subscribe to "info", so
it won't get the filenames.