Open animation controls window again after close
If you open an animations window, the button stores a reference to that
window. If you close the controls window, the button _still_ has a
reference to it, but it's closed. When the animation controls button is
clicked again, either focus and opn window, or delete the 'old_window'
property and re-open a new window.

The 'closed' property tells if the window is closed. I don't _think_
anything needs to be done to garbage collect the old window; at least,
it disappears from the task manager in Brave.