Static animation controls shouldn't be cleared
The animation control window adds some static controls and the rest of
the controls come from the webserver. The web server sends a "clear"
control command and we don't want to clear the static controls. We
could, but then we'd have to recreate them. Since the animation
processes aren't going to change them, we can just keep them around.

The static controls are:
- start / stop
- freeze / unfreeze
- timing display