Fix saving and loading
Some animations have state fields that shouldn't be saved, so I've added
the ability to wrap the fields in a {skip, _} tuple.

Also, when we load animations we need to add their Name => Pid entries
to the Animators map. (This presented as a bug when trying to stop an
animator by its name and not finding it in the map because it was loaded
from a file.)

Don't stop existing animators when a file is loaded. The user should be
able to layer multiple files on top of each other.