Loaded images each frame is ridiculous
Cache the images in a map so I can check if they're already loaded and
then just use the already loaded image.

Also, fixing the timing in the (very simple) image x animation: I use a
"cycle frames" number to limit the frame number to between 1 and 200
instead of using whatever the real frame number is (e.g. 6,328); this
allows me to take any frame and use it to figure out where in my back
and forth cycle the image should be. This allows me to calculate whether
I should be going left or right: if I'm in the first have of the cycle,
I'm going left, otherwise I'm going right; however, to calculate the
distance I'm going left, or right, I need _half_ the cycle: if I'm going
left, how far left am I going? For this I need to know which frame, of
the _first_ half of the cycle I'm on, so I use another modulo number:
frame-number modulo half-cycle.

e.g. If my cycle is 200, my frame is 6,328 then:
My cycle started at 6,200 and I'm on "cycle frame" 128.
Since 128 is > half of 200, I'm on the "right-to-left" portion of the
cycle. My "amount left/right" is 28, i.e. 6,328 rem 100.
This puts me at "right-most-point - 28", e.g. 600 - 28.

Using Promise to load the image

I used these MDN pages:
https://developer.mozilla.org/en-US/docs/Learn_web_development/Extensions/Async_JS/Promises
https://developer.mozilla.org/en-US/docs/Learn_web_development/Extensions/Async_JS/Implementing_a_promise-based_API
https://developer.mozilla.org/en-US/docs/Web/API/HTMLImageElement/complete

Explanation (for myself, since I'll forget this):

I'm using a promise to wait for an image to be loaded the first time we
see the image:

1| async function maybe_wait_image(drawCommand){
 |   const {cmd, src} = drawCommand;
 |   if(cmd == 'image' && !loadedImages.has(src)){
 |     const img = document.createElement('img');
 |     console.time('load image');
2|     await load_image(img, src);
 |     console.timeEnd('load image');
 |     loadedImages.set(src, img);
 |     drawCommand.img = img;
 |   }else if(cmd == 'image'){
 |     drawCommand.img = loadedImages.get(src);
 |   }
 | }

 | function load_image(img, url) {
3|   return new Promise((resolve, reject) => {
4|     img.onload = () => resolve(img);
5|     img.onerror = reject;
6|     img.src = url;
 |   });
 | }

1. 'async' means that this function waits on a promise. It is required.

2. Wait for the load_image/2 function to return. We could store the
result, but in this case (because we're passing references around
<GAG>), we don't need to.

  e.g. const result = await load_image(img, src)

  It feels like await should just be implied.

3. Create a Promise object which calls a function with a "resolve"
function and a "reject" function as arguments. These are functions that
_you_ call when you're ready. We return whatever the promise returns, in
this case the image, but we don't need to.

4. We attach the resolve function right to the img.onload property using
an "arrow function" so that we can pass the image itself: when the image
is loaded, resolve(img) will be called. The return of the resolve
function is what is returned from the promise, and which we simply return
back on line 2.

  I tried attached the bare resolve function to the onload property, but
  I got an error once about not being able to read the height of the image.

  I only tried once and then just set it back. I really don't care at
  this point.

5. attach the reject function directly to onerror. If there's an error,
... I don't know or care what happens.

6. Finally, we set the src on the image: this is what kicks off the
loading of the image by the browser. When the browser has loaded the
image the onload function will be called, which is the resolve function,
and our promise will return.