Set Z-index from controls
(Currently only works with images, since that is the only animation that
pays attention to the setting.)

This works like the "frame millis" setting, where the animator
gen_server eavesdrops on the value being set, and sends out the textbox
control after the animation module has sent its controls.

I changed the call to the animation modules to take Settings and
AnimationState instead of Frame and AnimationState: I changed all the
animators over to read the Frame from the "Settings" map. I could use a
better name but "State" is getting overloaded. Maybe "ServerState"? or
... I dunno. I don't want to get into "OO" terminolgy. "Generic State"?

I'm getting a flicker, not sure why.