Stream viewer challenge: have sidebar controls
A viewer of the Twitch stream challenged me to open the animation
controls in a sidebar on the main screen: the sidebar should show only
the controls for the animations whose controls have been opened the
latest: e.g. if you open the controls for "s_1", then "c_1", the "g_1",
only "c_1" and "g_1" will remain showing.

I split the code for the animation controls up so that I can call it
from a control window, or from the main window. The animation control
logic keeps everyone isolated from the main window. Each sidebar control
div has its own websocket back to the server.