Socket is set in context after buttons created
Since I create the static animation controls before the websocket is
open, I don't have access to the socket. I do, however, have access to
the context, which the socket is later added to. If I access
context.socket, then by the time the button is clicked the socket will
be there.