Clear canvas, clear elements, don't store (e) =>
Basically, fixed a bunch of bugs ... well, and starting to use the
"number" input type in place of text. I was actually going to use type
hints, but this just automatically limits to numbers. I lucked out in
that "number" is an input type just like "textbox", so I didn't need to
change anything.

- I was overloading clear for clearing child elements out of HTML
elements and clearing the canvas; this cleared up the mystery of why the
canvas wasn't getting cleared and animations were smearing: the 'clear'
function had been overridden.

- I was using element.remvoe instead of element.remvoeChild, hence the
need for a circuit breaker, because the element was just removing itself
(which is why I saw that it was removed from the elements tab in the
developer tools, duh)

- Apparently you can't return an arrow function. Switch to creating a
function inside another function and returning it worked fine. (I was
making an event handler closure that required outside variables.)

- Got rid of a bunch of "send_controls" code in ws_anim_channel that I
didn't end up needed. I think it's been there for a while.

TODO:
- z-index and frame millis aren't obeying the ".size = 10" setting;
they're intputs of type "number", like the other numbers; the only
difference I'm aware of, so far, is that they're sent by the animator
gen_server, not an animate module.
- I still need to add then channel frame_millis control so I can control
that from the web page
- Fix the circles2 animation