Anim ctrl buttons temp, clear all socket controls
All animation control buttons, i.e. to open up the control panel or
controls popup, are transient: as soon as a new websocket connects to
the channel, all sockets will receive a "clear animator names" command
and the buttons will be removed and recreated.

When a new socket joins the channel it needs to get the animator names
or the animator controls (the main page and the control panels/popups
all connect sockets to the shared channel). Whenever we send out
animator names or controls, we first send a clear so that no socket
connected to the channel gets duplicated animator names or controls.
So, we have to send the clear to _all_ sockets, not just a socket that
newly subscribed.

There are trade-offs to having all communication go through a single
channel despite having multiple front end processes (i.e. sockets) and
multiple back end processes (i.e. animators).

TODO:
- What happens when an animator dies and it has a control panel?
  Can animators die?
- Allow control panel/popups to connect to animators directly