Control panels/popups have single animator
Control panels and popups are the only web page entities that care about
animation controls, and they only care about the controls of one
animator. I've modied the 'control' subscription type to require the
name of an animator, and now only web sockets subscribed to
{AnimatorName, 'control'} will receive control updates for that
animator. These means all control panels and popups don't have to be
updated every time a new control panel or popup is created.

Send "clear controls" command from the animator gen_server, not from the
channel: this way when a control value is set, or 'send_controls' is
received, the animator can send the clear controls message.
'send_controls' occurs when a socket subscribes to the controls for an
animator.

The utils module now sends {AnimatorName, control} to the channel when
sending input controls so that the channel knows who to route them to
(i.e. any socket socket to controls for _that_ animator).

TODO:
- consider sending control value updates, instead of all new controls