Old circles2 lines should disappear
When lessening the number of lines drawn from the final cicrcle in the
"circles2" animation, if the number of lines is set to a lower number,
the "orphaned" lines in the channel's draw buffer need to be deleted.

Also, set min and max on the # of lines setting; something crashes if
the number of lines is less than 1.

TODO:
- Allow a 0 setting for number of lines and figure out what's breaking