Don't overwrite circle2 circle IDs with line IDs
I hard-coded the starting line ID for circles2 as 6, so as soon as I
created my sixth circle, it would be overwritten in the ETS table by the
first line who's frame was evenly divisible by #-of-frames. (Because
the lines get their ID based on the current frame, and loop around by
taking frame-# modulo #-of-lines.