Transform and Text animations
I created a transform "animation" (it updates if you change it) and a
text "animation" (same deal). I can add movement to them later on, but
for now, I'm trying to recreate this website:

  https://tinylittlemaggie.github.io/transformation-matrix-playground/

It's got an "F" that you can move around with transforms: flip, rotate,
skew, etc.

I created a transform "animation" that only modifies the canvas once and
then gets cached. It _can_ be modified to keep applying the transform
every frame, but right now I just want it to hold still.

Things to note:
- ctx.translate(x, y) is a *transform*
- ctx.scale(x, y) is a *transform*

If you use .setTransoform, then all previous transforms are reset, so
either use transform(...), or do your scale and translate *after* you
call setTransform(...).

If you transform the canvas, chances are you're going to draw out of the
visible canvas area; if you want to do rotations and such, then you need
to scale the canvas down so that you can see a larger area. Sound
confusing? It is: just because you scale the canvas, doesn't mean the
drawing area changes: it just means *where* the stuff you draw shows up.
e.g. if you scale the canvas by (0.5, 0.5), then everything you draw
will be in the top left of the canvas, but you can still draw outside of
that area; if you're canvas is 400x400, then you scale by (0.5, 0.5),
when you draw a 300x300 box at 0,0, it will actually be drawn as a
150x150 box; you could draw it at x = 400, y = 400, and it will actually
still be on the screen in the bottom right.

So, the trick that tinylittlemaggie.github.io does, is scale the canvase
down to 25%, then move it over to the top right corner so that when they
do the rotations and the 'F' moves "outside" of bounds, it still gets
drawn.

TODO:
- clear the _entire_ canvas, by taking the translate and scaling into
account (or by using -1000,-1000,3000,3000 (x,y,w,h) or something.