Add new carousel animator
Use Z-index to put boxes in the "front" or "back"

Since the trigonometric functions sine and cosine go from 1 to -1, we
multiply by half of whatever amount we're trying to animate, and then
add the other half to the result: if the value of sine/cosine is -1,
then we get (-1 * half + half) = 0, and if the value of sine/cosine is
1, then we get (1 * half + half) = full, where "full" is the full value.

We use cosine to determine the X coordinate because it's offset from
sine by 90 degress, making the boxes half-way across the screen when
they are at their largest (since size is determined by sine).

We added a "rise" element to the Y coordinate so that the boxes
would be lower down at the front, making it seeem like the boxes
are coming in at "feet level".