WIP: Capture web cam, send to server, send back
I want to capture images from a web cam on the browser's computer, send
it to the server to be convoluted, and then send it back to be drawn on
the canvas.

'async' does mean that a JavaScript function will run asynchronously:
the browser won't wait for calls to async functions to return: instead,
an async function will immediately return a Promise.

'await' will wait for a returned Promise to be resolved.

I had trouble with the <video> not being ready to a VideoFrame, so I
used the requestVideoFrameCallback function to call the capture function
when a new frame was ready; since I don't call requestVideoFrameCallback
again from within the callback function, we only get one frame. A
message from the server will kick off another frame capture when the
server is ready.

TODO:
- draw the data returned from the server onto the canvas
- do convolutions
- add controls for convolutions