Channel buffer resending identical 4mb images
The convolution animator requests an image from web page, and waits for
it. Once the animator receives an image, it convolutes it and then sends
it to the channel to be buffered. Once the channel gets the signal to
send the buffered draw commands to the webpage, it sends them all and
leaves them in the buffer: this is so that if an animator is slower than
the channel buffer send cycle, the last draw command of the animator
will be resent: in the case of the convolution animator, this means
image files left in the channel buffer are resent over and over.

This change adds a caching mechanism where commands in the channel
buffer can be sent once with a cache instruction for the web page, and
thereafter they are sent as "cached" commands with an ID; the web page
caches the draw command and then when it receives the "cached" command
with an ID, it pulls the draw command (e.g. containing 4mb of image
data) from the cache and draws that.

This isn't working well in the browser: the main thread can't seem to
keep up, even with the caching strategy.

TODO:
- use webworkers to offload some work off the main UI thread