Need to get logic off the main thread
Use web workers to offload work from the main thread:
- web socket communications
- animation buffering and drawing
- downloading images from the <video> element

This worked very well.

Layout:

main thread (i.e. with window) "main"
 - bg_socket.js web worker "socket"
   - bg_draw.js web worker "draw"
 - bg_send_image.js web worker "send_image"

When transferring an offline canvas to a web worker, make sure to add
the offline canvas to the list of "transfer" objects, both when you
transfer the offline canvase from main to socket, and from socket to
draw.

The docs say that both socket and draw should be able to call
requestAnimationFrame; however, I was only able to get it to work for
"socket".

  https://developer.mozilla.org/en-US/docs/Web/API/DedicatedWorkerGlobalScope/requestAnimationFrame

  ░ Calling the requestAnimationFrame() method requires the current worker
  ░ to have an associated owner window. That means that the current worker
  ░ must be created by window or by a dedicated worker that also has an
  ░ associated owner window.

  I'm not sure how to check that socket has "an associated owner window"
  or verify that draw doesn't.

The draw web worker can't access any DOM objects, so for any images
loaded from URL we'll need to send the data to the draw worker to be
cached.

TODO:
  - get all the animations working:
    - test video frames
  - send image data for images loaded from a URL to the draw web worker
    to be stored in a cache.
  - consider bundling draw messages from the server

Co-authored by: Terry Wilcox <terry.wilcox@mac.com>