Keep sketched out ideas for convolution matrix
I'm trying to work out how to create nine different copies of the image,
each with it's pixels aligned according to the center pixel. That way, I
can just take the head of each image binary, and those are the nine
pixels that correspond to the nine kernel values: I can multiply each
pixel by it's kernel value and add them together.

I'd rather do this than walk through one copy of the image, and for each
pixel figure out what the surrounding eight pixels are.