I want to use convolution kernels
I created code to make 9 different copies of the image where each copy
is spaced in such a way that a kernel can be applied to all bytes of
each image copy at a particlar index:

For example, we can apply the convolution kernel to the first 4 bytes of
each image: the first pixel, from the first image, will match the pixel
referenced by the top left value of the kernel; the first pixel from the
second image will match the pixel referenced by the top middle value of
the kernel.

In other words, instead of finding all the pixels referenced by the nine
values of the kernel in a single list, we create nine lists spaced out
so that the values for each kernel pixel position are aligned.

I'm not sure how to say this clearer.

If you imagine the kernel as a "window" sliding over the image, each
kernel value will see a different pixel; however, at the edges, for some
of the kernel values, there *is no* pixel and the value should be
assumed to be zero. So, since we don't have a 2D image to move the
kernel over, we can create nine images, and "slide" each of the nine
kernel values over it's own copy of the image, such that the pixels
being seen by each kernel value line up correctly.

Here's an imaginary 4 pixel x 4 pixel black and white image, with no
alpha, represented as a 2D grid of pixel values:

 1  2  3  4
 5  6  7  8
 9 10 11 12
13 14 15 16

  Here's a 3x3 kernel with variable place holders:

  a, b, c,
  d, e, f
  g, h, i

In order to apply this kernel to pixel 1,1:

  a doesn't point at anything (_)
  b -> _
  c -> _
  d -> _
  e -> 1
  f -> 2
  g -> _
  h -> 5
  i -> 6

  a, b, c, d, g don't point at anything.

If we lay out the image in 9 copies

a -> _,  _,  _,  _,  _,  1,  2,  3,  _,  5,  6,  7,  _,  9, 10, 11
b -> _,  _,  _,  _,  1,  2,  3,  4,  5,  6,  7,  8,  9, 10, 11, 12
c -> _,  _,  _,  _,  2,  3,  4,  _,  6,  7,  8,  _, 10, 11, 12,  _
d -> _,  1,  2,  3,  _,  5,  6,  7,  _,  9, 10, 11,  _, 13, 14, 15
e -> 1,  2,  3,  4,  5,  6,  7,  8,  9, 10, 11, 12, 13, 14, 15, 16 (identity)
f -> 2,  3,  4,  _,  6,  7,  8,  _, 10, 11, 12,  _, 14, 15, 16,  _
g -> _,  5,  6,  7,  _,  9, 10, 11,  _, 13, 14, 15,  _,  _,  _,  _
h -> 5,  6,  7,  8,  9, 10, 11, 12, 13, 14, 15, 16,  _,  _,  _,  _
i -> 6,  7,  8,  _, 10, 11, 12,  _, 13, 14, 15,  _,  _,  _,  _,  _

Now, if we compare the expected values that each kernel position will
"see", we can verify that each "seen" value matches up with the first
pixel value of each image.

Therefore, given all thse images, with their spacings, we can walk
through each column, grabbing one pixel from each row and matching it up
with the nine elements of the kernel.

So, instead of sliding one 3x3 kernel over a 2D image, we're sliding a
1x9 kernel over 9 1D images: each kernel position gets a bespoke copy of
the image with the pixels aligned.

TODO:
- Fix the black vertical banding: when values of 1 are used on the left
or right sides of the kernel, we get some black banding.

Here's some code to lay an image so it's easy to see:

  ░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░
  ░
  ░ 1> P = fun(P, I, [R, G, B, A | Rest]) ->
  ░            io:format("~3.. b: ~3.. b, ~3.. b, ~3.. b, ~3.. b~n",
  ░                      [I, R, G, B, A]),
  ░            P(P, I + 1, Rest);
  ░           (_, _, _) ->
  ░           ok
  ░         end.
  ░ #Fun<erl_eval.40.39164016>
  ░ 2> P(P, 1, I).
  ░
  ░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░

Example output: (Index, Red, Green, Blue, Alpha)

  ░   1:  13,   0,   0, 255
  ░   2:   4,   0,   0, 255
  ░   3:  51,   0,   3, 255
  ░   4:   0,   0,   0, 255
  ░   5:  14,   0,   0, 255
  ░   6:   4,   0,   0, 255
  ░   7:   9,   0,   0, 255
  ░   8:   0,   0,   0, 255
  ░   9:   9,   0,   0, 255
  ░  10:  14,   0,   0, 255
  ░  11:   1,   0,   0, 255