Hard-coded image width was wrong: 4 vs 160
I was getting black bands in the image because I was modifying the image
as if it was 4 pixels wide: the end of each 4-pixel "row" was a blank
pixel: I was seeing 160-pixel rows made up of 40 rows, each with a black
pixel on the right, concatenated together.

Also cleaned up the shifting logic with macros.
Erlang complains about a macro as a binary size specifier if the
macro is two levels deep: e.g.
-define(A, foo).
-define(B, ?A).

Co-authored by: Terry Wilcox <terry.wilcox@mac.com>